requirements.txt
ollama>=0.4.0
PyQt5>=5.15.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
psutil>=5.9.0
